# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VolleyPass API",
        "description": "API para el sistema de gestión de voleibol VolleyPass",
        "contact": {
            "name": "VolleyPass Support",
            "email": "admin@volleypass.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "VolleyPass API Server"
        }
    ],
    "paths": {
...
        "/api/public/tournaments/{id}/standings": {
            "get": {
                "tags": [
                    "Public Tournament Standings"
                ],
                "summary": "Obtener tabla de posiciones de un torneo",
                "operationId": "68fe36ee945432a6d1ad47e4310fb44c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del torneo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tabla de posiciones del torneo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "tournament_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tournament_name": {
                                                    "type": "string",
                                                    "example": "Copa Nacional 2024"
                                                },
                                                "format": {
                                                    "type": "string",
                                                    "example": "round_robin"
                                                },
                                                "groups": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "group_id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "group_name": {
                                                                "type": "string",
                                                                "example": "Grupo A"
                                                            },
                                                            "standings": {
                                                                "type": "array",
                                                                "items": {
                                                                    "properties": {
                                                                        "position": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "team_id": {
                                                                            "type": "integer",
                                                                            "example": 5
                                                                        },
                                                                        "team_name": {
                                                                            "type": "string",
                                                                            "example": "Equipo A"
                                                                        },
                                                                        "matches_played": {
                                                                            "type": "integer",
                                                                            "example": 3
                                                                        },
                                                                        "wins": {
                                                                            "type": "integer",
                                                                            "example": 2
                                                                        },
                                                                        "losses": {
                                                                            "type": "integer",
                                                                            "example": 1
                                                                        },
                                                                        "sets_won": {
                                                                            "type": "integer",
                                                                            "example": 6
                                                                        },
                                                                        "sets_lost": {
                                                                            "type": "integer",
                                                                            "example": 4
                                                                        },
                                                                        "points_for": {
                                                                            "type": "integer",
                                                                            "example": 150
                                                                        },
                                                                        "points_against": {
                                                                            "type": "integer",
                                                                            "example": 120
                                                                        },
                                                                        "table_points": {
                                                                            "type": "integer",
                                                                            "example": 7
                                                                        },
                                                                        "set_ratio": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 1.5
                                                                        },
                                                                        "point_ratio": {
                                                                            "type": "number",
                                                                            "format": "float",
                                                                            "example": 1.25
                                                                        }
                                                                    },
                                                                    "type": "object"
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "overall_standings": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "position": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "team_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "team_name": {
                                                                "type": "string",
                                                                "example": "Equipo A"
                                                            },
                                                            "group_name": {
                                                                "type": "string",
                                                                "example": "Grupo A"
                                                            },
                                                            "matches_played": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "wins": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "losses": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "table_points": {
                                                                "type": "integer",
                                                                "example": 7
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Torneo no encontrado"
                    }
                }
            }
        },
        "/api/public/tournaments/{id}/groups/{groupId}/standings": {
            "get": {
                "tags": [
                    "Public Tournament Standings"
                ],
                "summary": "Obtener tabla de posiciones de un grupo específico",
                "operationId": "b7eb4fb3d072a01f2f9968e26fcb2137",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del torneo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "groupId",
                        "in": "path",
                        "description": "ID del grupo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tabla de posiciones del grupo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "group_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "group_name": {
                                                    "type": "string",
                                                    "example": "Grupo A"
                                                },
                                                "tournament_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "tournament_name": {
                                                    "type": "string",
                                                    "example": "Copa Nacional 2024"
                                                },
                                                "standings": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "position": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "team_id": {
                                                                "type": "integer",
                                                                "example": 5
                                                            },
                                                            "team_name": {
                                                                "type": "string",
                                                                "example": "Equipo A"
                                                            },
                                                            "matches_played": {
                                                                "type": "integer",
                                                                "example": 3
                                                            },
                                                            "wins": {
                                                                "type": "integer",
                                                                "example": 2
                                                            },
                                                            "losses": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "sets_won": {
                                                                "type": "integer",
                                                                "example": 6
                                                            },
                                                            "sets_lost": {
                                                                "type": "integer",
                                                                "example": 4
                                                            },
                                                            "points_for": {
                                                                "type": "integer",
                                                                "example": 150
                                                            },
                                                            "points_against": {
                                                                "type": "integer",
                                                                "example": 120
                                                            },
                                                            "table_points": {
                                                                "type": "integer",
                                                                "example": 7
                                                            },
                                                            "set_ratio": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 1.5
                                                            },
                                                            "point_ratio": {
                                                                "type": "number",
                                                                "format": "float",
                                                                "example": 1.25
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Grupo o torneo no encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar el token de autenticación en el header Authorization: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}